Enhance backtesting framework and diagnostics
- Added new diagnostic fields in BacktestResult for tracking signal counts, PnL by pattern, and direction counts.
- Refined equity and drawdown tracking logic; introduced EquityPoint class.
- Updated BacktestRequest to include StrategyConfig.
- Streamlined BacktestService by removing IPatternService dependency.
- Improved pattern detection logic in PatternService and added new pattern types in PatternType.
- Enhanced SignalResult and SignalService for better signal extraction and handling.
- Updated TradeResult and TradeSimulator for improved trade execution tracking.
- Corrected configuration property naming in appsettings.Development.json.

+++ PatternDashboard/Models/Shared/TradeSimulator.cs
@@ -9,7 +9,7 @@ public class TradeSimulator : ITradeSimulator
     if (entryCandle == null || pattern == null)
       return null;
 
-      decimal entryPrice = (decimal)entryCandle.Open;
+      decimal entryPrice = entryCandle.Open;
       decimal stopLoss = pattern.Direction == "Buy"
     ? entryPrice - (decimal)(pattern.Config.StopLossPips * 0.0001)
     : entryPrice + (decimal)(pattern.Config.StopLossPips * 0.0001);
